+++ lib/src/other_utils/other_utils.dart
@@ -1,10 +1,14 @@
 /// Utilities for type conversion operations
 library;
 
+/// Utilities for handling object conversion.
 export 'convert_types.dart';
 
 /// Utilities for handling country codes and timezone operations
 export 'country_and_timezone_utils.dart';
+
+/// Utilities for handling function execution control
+export 'debouncer.dart';
 
 /// Implementation of a doubly linked list data structure
 export 'doubly_linked_list.dart';